requirements.txt
piexif==1.1.3
Pillow==9.5.0
